+++ examples/nextjs-rds-drizzle/packages/web/app/api/users/route.ts
@@ -1,0 +1,6 @@
+import { db, users } from "@/drizzle";
+
+export async function GET() {
+  const usersList = await db.select().from(users);
+  return Response.json({ usersList })
+}