+++ examples/nextjs-rds-drizzle/packages/web/sst-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="./.sst/types/index.ts" />