+++ packages/sst/src/constructs/SsrSite.ts
@@ -476,6 +476,10 @@ export interface SsrSiteProps {
       | "logRetention"
     > &
       Pick<FunctionProps, "copyFiles">;
+    /**
+     * @hidden Need to fix tsdoc generation to support the `Plan` type
+     */
+    transform?: (args: Plan) => void;
   };
 }
 
@@ -585,6 +589,7 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
     // Build app
     buildApp();
     const plan = this.plan(bucket);
+    transformPlan();
     validateTimeout();
 
     // Create CloudFront
@@ -707,6 +712,10 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
       });
     }
 
+    function transformPlan() {
+      cdk?.transform?.(plan);
+    }
+
     function createServerFunctionForDev() {
       const role = new Role(self, "ServerFunctionRole", {
         assumedBy: new CompositePrincipal(
@@ -748,10 +757,7 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
       // Create warmer function
       const warmer = new CdkFunction(self, "WarmerFunction", {
         description: "SSR warmer",
-        code: Code.fromAsset(
-          plan.warmerConfig?.function ??
-            path.join(__dirname, "../support/ssr-warmer")
-        ),
+        code: Code.fromAsset(path.join(__dirname, "../support/ssr-warmer")),
         runtime: Runtime.NODEJS_18_X,
         handler: "index.handler",
         timeout: CdkDuration.minutes(15),
@@ -765,8 +771,7 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
 
       // Create cron job
       new Rule(self, "WarmerRule", {
-        schedule:
-          plan.warmerConfig?.schedule ?? Schedule.rate(CdkDuration.minutes(5)),
+        schedule: Schedule.rate(CdkDuration.minutes(5)),
         targets: [new LambdaFunction(warmer, { retryAttempts: 0 })],
       });
 
@@ -1015,7 +1020,7 @@ function handler(event) {
         ...cdk?.server,
         streaming: props.streaming,
         injections: [
-          ...(warm ? [useServerFunctionWarmingInjection()] : []),
+          ...(warm ? [useServerFunctionWarmingInjection(props.streaming)] : []),
           ...(props.injections || []),
         ],
         prefetchSecrets: regional?.prefetchSecrets,
@@ -1235,7 +1240,7 @@ function handler(event) {
     }
 
     function useServerBehaviorCachePolicy() {
-      const allowedHeaders = plan.cachePolicyAllowedHeaders || [];
+      const allowedHeaders = plan.serverCachePolicy?.allowedHeaders ?? [];
       singletonCachePolicy =
         singletonCachePolicy ??
         new CachePolicy(
@@ -1258,15 +1263,24 @@ function handler(event) {
       return singletonOriginRequestPolicy;
     }
 
-    function useServerFunctionWarmingInjection() {
-      return `
-if (event.type === "warmer") {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve({ serverId: "server-" + Math.random().toString(36).slice(2, 8) });
-    }, event.delay);
-  });
-}`;
+    function useServerFunctionWarmingInjection(streaming?: boolean) {
+      return [
+        `if (event.type === "warmer") {`,
+        `  const p = new Promise((resolve) => {`,
+        `    setTimeout(() => {`,
+        `      resolve({ serverId: "server-" + Math.random().toString(36).slice(2, 8) });`,
+        `    }, event.delay);`,
+        `  });`,
+        ...(streaming
+          ? [
+              `  const response = await p;`,
+              `  responseStream.write(JSON.stringify(response));`,
+              `  responseStream.end();`,
+              `  return;`,
+            ]
+          : [`  return p;`]),
+        `}`,
+      ].join("\n");
     }
 
     function getS3FileOptions(copy: S3OriginConfig["copy"]) {
@@ -1582,12 +1596,10 @@ if (event.type === "warmer") {
       edgeFunction?: keyof EdgeFunctions;
     }[];
     errorResponses?: ErrorResponse[];
-    cachePolicyAllowedHeaders?: string[];
-    buildId?: string;
-    warmerConfig?: {
-      function: string;
-      schedule?: Schedule;
+    serverCachePolicy?: {
+      allowedHeaders?: string[];
     };
+    buildId?: string;
   }) {
     return input;
   }