+++ packages/sst/CHANGELOG.md
@@ -1,5 +1,83 @@
 # @serverless-stack/cli2
 
+## 2.40.3
+
+### Patch Changes
+
+- [`e9847fb13`](https://github.com/sst/sst/commit/e9847fb1399ea7e1af0a2bd4130339deabf49a8a) Thanks [@fwang](https://github.com/fwang)! - Cli: fix deploy status event not shown
+
+## 2.40.2
+
+## 2.40.1
+
+## 2.40.0
+
+### Minor Changes
+
+- [#3644](https://github.com/sst/sst/pull/3644) [`1ca805d26`](https://github.com/sst/sst/commit/1ca805d2673f25dd8790422ac1154fb3da26c91c) Thanks [@fwang](https://github.com/fwang)! - Update CDK to v2.124.0
+
+### Patch Changes
+
+- [#3586](https://github.com/sst/sst/pull/3586) [`729e5fe15`](https://github.com/sst/sst/commit/729e5fe158a01ce38da2ebbf92c214b05242367a) Thanks [@BLucky-gh](https://github.com/BLucky-gh)! - Function: Add cache options for python bundling
+
+## 2.39.13
+
+### Patch Changes
+
+- [`b4a826ae3`](https://github.com/sst/sst/commit/b4a826ae312bed164365c561cfbfa221092373a2) Thanks [@fwang](https://github.com/fwang)! - SsrSite: handle s3 upload rate limit error
+
+## 2.39.12
+
+### Patch Changes
+
+- [`ecba06be8`](https://github.com/sst/sst/commit/ecba06be80e99a8b064cf65718b6a188537e640e) Thanks [@fwang](https://github.com/fwang)! - Function: fix build hook run after CDK creates artifacts folder for container runtime
+
+## 2.39.11
+
+### Patch Changes
+
+- [`4f2302921`](https://github.com/sst/sst/commit/4f23029218869fc25a7ec9c49cfe44349c07ced7) Thanks [@fwang](https://github.com/fwang)! - NextjsSite: support per-route logging for Route Handlers
+
+## 2.39.10
+
+### Patch Changes
+
+- [#3630](https://github.com/sst/sst/pull/3630) [`953a1ac70`](https://github.com/sst/sst/commit/953a1ac7026d242c04b7ef3220236a2119ae0599) Thanks [@mathisobadia](https://github.com/mathisobadia)! - Job: Allow setting nodejs version of container image
+
+- [`c8ac970af`](https://github.com/sst/sst/commit/c8ac970af8634f0304af87066a622396dd04a3cb) Thanks [@fwang](https://github.com/fwang)! - Bootstrap: bump version to trigger CDK custom resource runtime update
+
+- [#3516](https://github.com/sst/sst/pull/3516) [`0d9cf1172`](https://github.com/sst/sst/commit/0d9cf1172638b3f92634982bac4155a04d5995de) Thanks [@morficus](https://github.com/morficus)! - NextjsSite: new option to specify custom build output path
+
+## 2.39.9
+
+### Patch Changes
+
+- [`83d013b35`](https://github.com/sst/sst/commit/83d013b354ac9bf40031642b09c9ae0e18e007e2) Thanks [@fwang](https://github.com/fwang)! - NextjsSite: fix warming when streaming is enabled
+
+## 2.39.8
+
+### Patch Changes
+
+- [`b29ce15fa`](https://github.com/sst/sst/commit/b29ce15fa94305cc50e4a40b95476cf8bd316bca) Thanks [@fwang](https://github.com/fwang)! - SsrSite: support customization via "cdk.transform" prop
+
+## 2.39.7
+
+### Patch Changes
+
+- [`cfb9f5814`](https://github.com/sst/sst/commit/cfb9f5814c828c26339ca0bb68f3f896e2375ff6) Thanks [@fwang](https://github.com/fwang)! - Revert adding @smithy/util-endpoints
+
+## 2.39.6
+
+### Patch Changes
+
+- [`ff3abe365`](https://github.com/sst/sst/commit/ff3abe36516b0e110978390178fe9c17fbac294e) Thanks [@thdxr](https://github.com/thdxr)! - fix: removed broken ion code
+
+## 2.39.5
+
+### Patch Changes
+
+- [`daec2ff64`](https://github.com/sst/sst/commit/daec2ff64b275475302e52c236f9073f27f10c94) Thanks [@fwang](https://github.com/fwang)! - Revert adding @smithy/util-endpoints
+
 ## 2.39.4
 
 ### Patch Changes